compounder-specific hooks
