+++ src/hooks/data/compounders/useListCompounders.ts
@@ -1,21 +1,16 @@
 import { useContractRead } from "wagmi"
 
-import { VaultProps } from "@/lib/types"
+import { VaultType } from "@/lib/types"
 import { useApiCompounderVaults, useApiTokenVaults } from "@/hooks/api"
-import { useVaultTokens } from "@/hooks/data"
 import useRegistryContract from "@/hooks/useRegistryContract"
 import {
   useIsCurveCompounder,
   useIsTokenCompounder,
 } from "@/hooks/useVaultTypes"
 
-export default function useVaultName({ asset, type }: VaultProps) {
+export function useListCompounders({ type }: { type: VaultType }) {
   const isCurve = useIsCurveCompounder(type)
   const isToken = useIsTokenCompounder(type)
-  const { data: vaultTokens } = useVaultTokens({
-    asset,
-    type,
-  })
 
   // Preferred: API request
   const apiCompounderQuery = useApiCompounderVaults({ type })
@@ -25,30 +20,30 @@ export default function useVaultName({ asset, type }: VaultProps) {
   const registryQuery = useContractRead({
     ...useRegistryContract(),
     functionName: isToken
-      ? "getTokenCompounderName"
+      ? "getTokenCompoundersList"
       : isCurve
-      ? "getCurveCompounderName"
-      : "getBalancerCompounderName",
-    args: [asset ?? "0x"],
+      ? "getCurveCompoundersList"
+      : "getBalancerCompoundersList",
     enabled: apiCompounderQuery.isError || apiTokenQuery.isError,
   })
+
   // Prioritize API response until it has errored
   if (!apiCompounderQuery.isError && !isToken) {
     return {
       ...apiCompounderQuery,
-      data: apiCompounderQuery.data?.find(
-        (p) => p.token.ybToken.address === vaultTokens.ybTokenAddress
-      )?.poolName,
+      data: apiCompounderQuery.data?.map((p) => p.token.LPtoken?.address),
     }
   }
+
   if (!apiTokenQuery.isError && isToken) {
     return {
       ...apiTokenQuery,
-      data: apiTokenQuery.data?.find(
-        (p) => p.token.ybToken.address === vaultTokens.ybTokenAddress
-      )?.vaultName,
+      data: apiTokenQuery.data?.map((p) => p.token.baseAsset.address),
     }
   }
-  // Fallback to contract data after failure
-  return registryQuery
+  const filterRegistryQuery = {
+    ...registryQuery,
+    data: registryQuery.data,
+  }
+  return filterRegistryQuery
 }
