+++ src/hooks/data/concentrators/index.ts
@@ -1,3 +1,4 @@
+export * from "./useConcentrator"
 export * from "./useConcentratorRewards"
 export * from "./useConcentratorTargetAssets"
-export * from "./useConcentratorVaults"
+export * from "./useListConcentrators"
