+++ src/hooks/data/index.ts
@@ -1,4 +1,3 @@
-export { default as useVaultAddresses } from "./useVaultAddresses"
 export {
   useVaultAuraApr,
   useVaultBalApr,
@@ -11,9 +10,7 @@ export {
   useVaultTotalApr,
 } from "./useVaultApy"
 export { default as useVaultDepositedAssets } from "./useVaultDepositedAssets"
-export { default as useVaultName } from "./useVaultName"
 export { default as useVaultPlatformFee } from "./useVaultPlatformFee"
 export { default as useVaultPoolId } from "./useVaultPoolId"
-export { default as useVaultTokens } from "./useVaultTokens"
 export { default as useVaultTvl } from "./useVaultTvl"
 export { default as useVaultWithdrawFee } from "./useVaultWithdrawFee"
