+++ src/hooks/data/useVaultPoolId.ts
@@ -1,11 +1,11 @@
 import { VaultProps } from "@/lib/types"
 import { useApiCompounderVaults, useApiTokenVaults } from "@/hooks/api"
-import { useVaultTokens } from "@/hooks/data"
+import { useCompounder } from "@/hooks/data/compounders"
 import { useIsTokenCompounder } from "@/hooks/useVaultTypes"
 
 export default function useVaultPoolId({ asset, type }: VaultProps) {
   const isToken = useIsTokenCompounder(type)
-  const { data: vaultTokens } = useVaultTokens({
+  const { data: vaultTokens } = useCompounder({
     asset,
     type,
   })
